LayoutTests/platform/ios-simulator/imported/w3c/web-platform-tests/css/css-lists/list-and-writing-mode-001-expected.txt


FAIL list and writing-mode assert_equals: the height of li should be 45px, and no extra line generated expected 45 but got 65
a b c
